fix: integer overflow in envelope.go

+++ envelope.go
@@ -57,7 +57,7 @@ func (e *envelope) IsSet(flag uint8) bool {
 // Read implements [io.Reader].
 func (e *envelope) Read(data []byte) (readN int, err error) {
 	if e.offset < 5 {
-		prefix, err := makeEnvelopePrefix(e.Flags, e.Data.Len())
+		prefix, err := makeEnvelopePrefix(e.Flags, uint32(e.Data.Len()))
 		if err != nil {
 			return 0, err
 		}
@@ -80,7 +80,7 @@ func (e *envelope) Read(data []byte) (readN int, err error) {
 // WriteTo implements [io.WriterTo].
 func (e *envelope) WriteTo(dst io.Writer) (wroteN int64, err error) {
 	if e.offset < 5 {
-		prefix, err := makeEnvelopePrefix(e.Flags, e.Data.Len())
+		prefix, err := makeEnvelopePrefix(e.Flags, uint32(e.Data.Len()))
 		if err != nil {
 			return 0, err
 		}
@@ -374,7 +374,7 @@ func (r *envelopeReader) Read(env *envelope) *Error {
 	return nil
 }
 
-func makeEnvelopePrefix(flags uint8, size int) ([5]byte, error) {
+func makeEnvelopePrefix(flags uint8, size uint32) ([5]byte, error) {
 	if size < 0 || size > math.MaxUint32 {
 		return [5]byte{}, fmt.Errorf("connect.makeEnvelopePrefix: size %d out of bounds", size)
 	}
